requirements.txt
streamlit==0.82.0
Pillow==8.2.0
urllib3==1.24.3
numpy==1.19.5
pandas==1.1.5
pip install fastai==1.0.61